PageRank/requirements.txt
requests==2.31.0
networkx==3.2.1
matplotlib==3.8.2
python-dotenv==1.0.1